+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Fixed an error that occurred when eager loading Neo block content
+- Fixed styling issues with block action menus
 
 ## 5.0.0-beta.8 - 2024-04-29